Adversarial review batch 4: comma chains structured, puns adjudicated
Expression-lens remainders, all verified MATCH + CHECK-GREEN:
- ActATTACK.c: both decompiler comma chains rewritten as structured
  nested ifs (byte-identical), and the shifted_mid register pun proved
  removable outright -- UpdateMotion takes saved_mid directly and the
  status write is a plain `saved_mid >> 8`; the dead shifted_mid and
  updated locals are gone too.
- DefaultActionHumanoid.c: the width fixed-point pun loses its
  redundant (s32) cast; the remaining (u16) view is the retail lhu
  access width (noted inline).
- AttackControl.c: the *(u32 *)&map.vector & 0xffff00ff word test is
  byte-required (one lw over four bytes, mask skips `direct`); noted
  what it actually tests.
- ActSTICKON.c / CVAupdate.c: the two shared high-half puns are real
  (one sll feeds two different shift-extracts: pad word + PADLup bit;
  CVA x-slot low byte + think-index high byte); annotated, kept.

+++ src/main.exe/DefaultActionHumanoid.c
@@ -303,7 +303,7 @@ short DefaultActionHumanoid(Humanoid *human)
                 }
                 else
                 {
-                    i = ((s32)((u16)human->width << 16)) >> 18;
+                    i = ((u16)human->width << 16) >> 18; /* width / 4; the u16 view is the retail lhu access width */
                     xx = RefrectMove[direction][0] * i;
                     zz = RefrectMove[direction][1] * i;
                 }
